requirements.txt
requests==2.32.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.7
uvicorn==0.30.1
fastapi==0.111.0
pytest==8.2.2